refactor: create structures/templates folder for BaseStructure
